Move handler logic to the internal folder
