--- cmd/payload-tracker-consumer/handler.go
@@ -1,5 +1,0 @@
-package main
-
-func onMessage() {
-	// Add Message handler here
-}
