+++ internal/handlers/handler.go
@@ -1,0 +1,5 @@
+package handlers
+
+func onMessage() {
+	// TODO: Add Message handler here
+}
